LayoutTests/imported/w3c/web-platform-tests/css/css-shadow-parts/simple-inline-expected.txt

The following text should be green: 

FAIL Part in selected host is styled assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
